requirements.txt
ado-py==0.0.7
cchardet==2.1.7
aiodns==2.0.0
aiogram==2.11.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.4
idna==2.10
multidict==5.0.2
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
python-dotenv==0.15.0
pytz==2020.4
toml==0.10.2
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
yarl==1.6.3
tqdm==4.56.0
pillow==8.1.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cpu
torchvision==0.8.2+cpu
torchfile==0.1.0
requests==2.25.1
googledrivedownloader==0.4
wget==3.2

